+++ src/mapgen/mapgen.c
@@ -29,7 +29,8 @@
 #include "mapgen.h"
 #include "ilist.h"
 #include "ilist_test.h"
-#include "rlist.h"
+#include "vectors/roads.h"
+#include "vectors/tiles.h"
 #include "sfc32_test.h"
 #include "random.h"
 #include "random_test.h"
