+++ src/vectors/tiles.h
@@ -3,15 +3,8 @@
 // Copyright (c) 2022 by the OlyTag authors.
 // Please see the LICENSE file in the root directory of this repository for further information.
 
-#ifndef OLYTAG_RLIST_H
-#define OLYTAG_RLIST_H
-
-// a collection of reallocation lists
-
-typedef struct road **road_list;
-
-int road_list_len(road_list l);
-void road_list_append(road_list *l, struct road *n);
+#ifndef OLYTAG_VECTOR_TILES_H
+#define OLYTAG_VECTOR_TILES_H
 
 typedef struct tile **tiles_list;
 
@@ -21,4 +14,4 @@ tiles_list tiles_list_copy(tiles_list l);
 void tiles_list_reclaim(tiles_list *l);
 void tiles_list_scramble(tiles_list l);
 
-#endif //OLYTAG_RLIST_H
+#endif //OLYTAG_VECTOR_TILES_H
